notification fix, added chronological timeline for scheduled posts

+++ app/src/main/java/org/nuclearfog/twidda/backend/async/UserlistAction.java
@@ -8,7 +8,6 @@ import androidx.annotation.Nullable;
 import org.nuclearfog.twidda.backend.api.Connection;
 import org.nuclearfog.twidda.backend.api.ConnectionException;
 import org.nuclearfog.twidda.backend.api.ConnectionManager;
-import org.nuclearfog.twidda.model.UserList;
 
 /**
  * async task to load list information and take action to the list
@@ -30,21 +29,14 @@ public class UserlistAction extends AsyncExecutor<UserlistAction.Param, Userlist
 	@Override
 	protected Result doInBackground(@NonNull Param param) {
 		try {
-			switch (param.mode) {
-				case Param.LOAD:
-					UserList result = connection.getUserlist(param.id);
-					return new Result(Result.LOAD, param.id, result, null);
-
-				case Param.DELETE:
-					connection.deleteUserlist(param.id);
-					return new Result(Result.DELETE, param.id, null, null);
-
-				default:
-					return null;
+			if (param.mode == Param.DELETE) {
+				connection.deleteUserlist(param.id);
+				return new Result(Result.DELETE, param.id, null);
 			}
 		} catch (ConnectionException exception) {
-			return new Result(Result.ERROR, param.id, null, exception);
+			return new Result(Result.ERROR, param.id, exception);
 		}
+		return null;
 	}
 
 	/**
@@ -52,14 +44,17 @@ public class UserlistAction extends AsyncExecutor<UserlistAction.Param, Userlist
 	 */
 	public static class Param {
 
-		public static final int LOAD = 1;
 		public static final int DELETE = 4;
 
 		final int mode;
 		final long id;
 
-		public Param(int mode, long id) {
-			this.mode = mode;
+		/**
+		 * @param action action to apply on the userlist {@link #DELETE}
+		 * @param id     userlist ID
+		 */
+		public Param(int action, long id) {
+			this.mode = action;
 			this.id = id;
 		}
 	}
@@ -69,21 +64,22 @@ public class UserlistAction extends AsyncExecutor<UserlistAction.Param, Userlist
 	 */
 	public static class Result {
 
-		public static final int LOAD = 5;
 		public static final int DELETE = 8;
 		public static final int ERROR = -1;
 
-		public final int mode;
+		public final int action;
 		public final long id;
-		@Nullable
-		public final UserList userlist;
 		@Nullable
 		public final ConnectionException exception;
 
-		Result(int mode, long id, @Nullable UserList userlist, @Nullable ConnectionException exception) {
-			this.userlist = userlist;
+		/**
+		 * @param action    action to apply on the userlist {@link #DELETE}
+		 * @param id        userlist ID
+		 * @param exception not null if an error occured
+		 */
+		Result(int action, long id, @Nullable ConnectionException exception) {
 			this.exception = exception;
-			this.mode = mode;
+			this.action = action;
 			this.id = id;
 		}
 	}
